Update patch set 1
Patch Set 1:

It is, but ...

It would (luckily!) be testing a part of a deprecated API we don't actually use. AFAICT, we are currently lacking tests for much more important parts of this API.

There is a SharedBufferTest.cpp in a somewhat odd place, but it's extremely limited, and I'm not sure how to run it, or if it is run automatically. There is also a Vector test, which tests it somewhat indirectly.

Maybe file a bug to fix SharedBuffer testing?

Patch-set: 1