Update patch set 1
Patch Set 1:

> ah right, yes.. sorry about that. compiling all tests into a single
 > executable (and then using --gtest_filter) seems to be the norm on
 > Android. oddly, the opposite is the norm in google3.

i think you can blame me for that. if we had blaze, i'd have been more inclined to follow the google3 convention. given that we don't, this makes "run all the tests" and "run all the tests with <pattern> in the name" easy and convenient, which makes it more likely to actually happen.

/me worries that the TreeHugger folks are wandering off in the opposite direction of complex configuration in their own custom DSL...

Patch-set: 1