Update patch set 1
Patch Set 1:

I talked with cferris offline to clarify that this patchstack deals with the ABI issues, it's just a question of how we want to manage the UAPI-facing definitions.

For now let's open-code both sets of ioctls, which this patchstack does already.  We can fix this up later when the new UAPI headers land in bionic.

Patch-set: 1