ueventd: replace strdup() in parse_config()
This commit replaces the strdup() with std::vector<char> to allow for
the copied string to be freed when it goes out of scope instead of
relying on the C strdup() and free() functions.

It's a fix similar to change Id0a5f711e33363082ba201afda6b26043998cb1c.

v2: duplicate args[1] to avoid pointing to memory of the copied string
which will be freed.

Change-Id: I4dbf543de6880537b419e6c2344ee6a40ce378f1