Bring back adb hell command
This wonderfull adb hell command has been forgotten during the current
adb refactoring. This commit brings it back.

Here we handle both "shell" and "hell" commands on the same block and
wrap around the adb_shell() call with the colorization characters
in case of "hell" command.

I tried to keep the way the parsing was done previously on the project.

Change-Id: Ib61408de38bf4190d03d64820c1e3057db0cd968
Signed-off-by: Eyal Lezmy <eyal.lezmy@gmail.com>