Update patch set 5
Patch Set 5:

I understand the copy is unnecessary and I had also originally done this, I think it'd be better for consistency with init_parser.cpp and conformance to C++ to have the copy.  The uevent files aren't super large, so it's not likely to have a big performance penalty.

I've added Yabin too, who helped review the init_parser.cpp change for his opinion.

Patch-set: 5