ueventd: notify dev node creation via property set event
Let ueventd set a property when a dev node is created so that init
may respond to it.

One needs to append "notify" to interested dev nodes in ueventd.rc
to enable the notification for those nodes.
The name of the property set by ueventd follows the format of
"add:<dev_path>" and its value is set to "<major>:<minor>".

Bug: 31586295
Test: WIP
Change-Id: Ibb8924036371b4c014ae0af8a69bba51afadf086