Update patch set 1
Patch Set 1:

The original patch was reverted because on some devices, adbd is confronted with an IO error just after opening the control file. And if we close the control file after IO error, then there will be an endless loop of "open control file -> io error -> open control file ...". Unless we fix that kernel or hardware bug, I am afraid we can't move on with this patch. Do you have any tests or procedures to reproduce the adb offline issue, so we can reconsider the importance of this patch?

Patch-set: 1