Update patch set 2
Patch Set 2:

SELinux domain transitions are not allowed when executing files from a nosuid mounted partition, since it's a sign that the partition isn't trusted to have executables which changes capabilities. SELinux domain transitions are very similar to executing a setuid binary, in that the permissions associated with a process can change.

Mounting /system nosuid also wouldn't work for /system/bin/run-as, since that file has file capabilities AND is expected to be called from a non-privileged process.

Because of /system/bin/run-as, we couldn't remove the existing android_filesystem_config.h mechanism we have today. This patch can't address the /system/bin/run-as use case executed from the unprivileged shell user.

Ignoring the debate over which technique is better (ambient capabilities vs file based capabilities), I hope all of us agree that configuration of process capabilities should be as simple and uniform as possible. Right now we already have two different ways to manage capabilities in processes

  1) file capabilities
  2) start as root and drop privileges (simplified somewhat via libminijail)

adding a third way increases the amount of code we need to maintain as well as introduces user confusion about the correct approach.

Or, quoting xkcd: https://xkcd.com/927/

In addition, this would also require our partners to backport the relevant kernel patches to their trees, so this would probably require some kind of CTS test to allow us to depend on it. Probably not a huge deal, but something we should be aware of.

Patch-set: 2
Label: Verified=0