Update patch set 2
Patch Set 2:

> I don't see why the AID_NET_RAW and AID_NET_ADMIN can't die and
 > fscaps  used in it's place. netadmin and netraw are not used in
 > system/core/rootdir/init.rc and are only present under the device
 > tree. All uses are on filesystems that support fscaps.
 > 
 > # repeat as needed
 > aosp/device$ find . -name init\.*\.rc | xargs isearch.py search
 > --tidy --section service --group net_raw
 > service(./lge/bullhead/init.bullhead.rc : 279): qti
 > /system/vendor/bin/qti
 > group : radio net_raw diag usb net_admin
 > 
 > <snip>
 > 
 > I checked my internal trees and its consistent with that above
 > statement.

Because if any of those services are launching helpers that also need NET_ADMIN or NET_RAW, then we have to add file caps for those binaries too. And that is cumbersome and error prone.

File caps don't inherit across execve(2). GIDs and ambient caps do. Let's replace GIDs with ambient caps, and allow folks to use ambient caps for other things too, which is gonna help because it's a lot easier to configure ambient caps in your init.rc file than to have to go find wherever file caps are defined.

I challenge the notion that having ambient caps available is gonna cause folks to do the wrong thing. If anything, it's gonna make it easier for folks to use capabilities: init will always configure things the right way, and it will allow you to configure capabilities for *process trees*, something that's currently impossible to do with file caps -- you can fake it by setting things up for a bunch of binaries, but you're still faking it. You could imagine a future in which ambient caps + seccomp allow us to configure "runtime environments" for process trees where system access can be configured in a significantly less involved way than with the current file caps + SELinux way.

Ambient caps is the natural (because of execve(2) inheritance) replacement for AID_*, and therefore allows for easier conversion, and faster dropping of those patches.

Patch-set: 2