libcutils: add android_get_control_file()
Solve one more issue where privilege is required to open a file and
we do not want to grant such to the service. The file's descriptor
is placed into the environment as "ANDROID_FILE_<path>" where . and /
in the <path> are replaced with _ and this function picks this up.

Added definition ANDROID_FILE_ENV_PREFIX for "ANDROID_FILE_" in
a new include <cutils/files.h>

Test: compile
Bug: 32450474
Change-Id: I2d0310a1727f1e393a00d9fc7e6cf5d028f27905