Update patch set 3
Patch Set 3:

(7 comments)

A few nits inline, but generally looks good.

It'd be great to combine this with SocketInfo somehow since they are very similar.  A lot of the CreateFiles()/CreateSockets(), PublishFile()/PublishSocket() functions should probably be a part of the SocketInfo and FileInfo classes.  Maybe if that's done first then it would be more clear how to combine them.

Patch-set: 3
Label: Verified=0