Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > (The royal) we are working on moving as much as possible from
 > init.rc into individual rc files located at /system/etc/init/

That's nice, where is the docs on configuring this? Should it just be
done with PRODUCT_COPY_FILES or is there a better way?
Does init automatically search there or does it need to be imported
somehow?

 > directory. As such, the only content that should reside in init.rc
 > will be about bringup. There is some /proc, /sys and /dev
 > preparations that can not be helped by this. Also all new products
 > should be designed A/B, which means that /system is mounted
 > immediately and then init is run. There is not much need for us to
 > add late bindings as a result.

Iv'e heard about this, yet to see any device like this, is hikey or the
new pixel like this?

Patch-set: 2