Update patch set 2
Patch Set 2:

> That's nice, where is the docs on configuring this? Should it just be
done with PRODUCT_COPY_FILES or is there a better way?
Does init automatically search there or does it need to be imported
somehow?

readme.txt in this same directory describes the LOCAL_INIT_RC build system macro that does this.

I don't have a strong opinion on having the late binding or not, but I'd prefer the late binding to happen in SetProcessAttributes() in all cases rather than only the cases where the lookup fails at the time of parsing.
There isn't much downside that I can see to doing this, besides that it would keep the user and group strings around for longer, but it would be cleaner code; the Parse* functions aren't meant to be called besides when parsing the service entries.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0