liblog: logprint supports number of seconds time event field
Add s to report time in seconds.  The time could be a period, duration
or monotonic, expanded to seconds, minutes, hours and days.  gTest is
also a demonstration of dynamic tag allocation.

Looking to the future, the audio media code has binary content similar
to the binary events structures Android logging uses and they have
a definition of a field in their internal binary logging, so may be
of use if we were to unify the logs.

Test: gTest logcat-unit-tests --gtest_filter=*.descriptive
Bug: 31456426
Change-Id: I262c03775983b3bc7b1b00227ce2bb2b0f357bec