liblog: logprint supports number of seconds time event field
Add s to report time in seconds.  The time could be a period, duration
or monotonic, expanded to seconds, minutes, hours and days.  Retry on
-EBUSY for all descriptive tests to survive heavy DOS load.  gTest is
also a demonstration of dynamic tag allocation.  Media has such a
definition of a field in their internal binary logging, so may be of
use in the future.

Test: gTest logcat-unit-tests --gtest_filter=*.descriptive
Bug: 31456426
Change-Id: I262c03775983b3bc7b1b00227ce2bb2b0f357bec