init.rc: move debugfs restorecon earlier
In frameworks/native commit
* f920891b71cbd33c96e7e44231524ee0b8e83760 (internal master)
* 23d030256136620102f9da216d2f3758ac1c13f3 (AOSP)

the chown/chmod/write calls were moved from the "on boot" section to the
"on fs" section, so that tracing would be available earlier in the boot
cycle. However, at "on boot", a restorecon has not yet been run on
/sys/kernel/debug, which means those debug files retain their same
SELinux label. However, init is not permitted to write to generically
labeled debugfs files, which causes the following write to fail:

  # Tracing disabled by default
  write /sys/kernel/debug/tracing/tracing_on 0

and results in the following dmesg error on boot:

  [   18.554609] audit: type=1400 audit(11012372.736:8): avc:  denied  {
  write } for  pid=1 comm="init" name="tracing_on" dev="debugfs" ino=7253
  scontext=u:r:init:s0 tcontext=u:object_r:debugfs:s0 tclass=file
  permissive=0
  [   18.554639] init: write_file: Unable to open
  '/sys/kernel/debug/tracing/tracing_on': Permission denied

Move the restorecon earlier in boot to accomodate the atrace.rc change.

This change relies on the fact that most device *.rc files mount debugfs
in early-init, and that init.rc's "on fs" section is executed prior to
aboot.rc's "on fs" section.

There's more cleanup that can occur here, such as removing individual
debugfs mounts from device-specific *.rc files, and having a consistent
time which debugfs mounts occur. However, that's left for another day.

Test: device boots
Test: /sys/kernel/debug files are labeled properly
Test: error messages above are no longer present
Change-Id: I09ba165beb13399262f898cddb713bb2592e080d