Update patch set 1
Patch Set 1: Code-Review+2

(3 comments)

Much improved, although I do wonder if DescriptorInfo might be too much abstraction. With this change, there's quite a bit of differences between sockets and files, and I'm not sure the code-reuse advantages are so great as to compensate for the additional complexity from this abstraction.

Anyway, it could be addressed in a future change if you decide to fix it.

Patch-set: 1
Label: Code-Review=+2