capabilities: Get last valid capability at runtime.
Parntners have expressed interest in using the 'capabilities' keyword
in init, so make the code more resilient against older kernels.

This implementation gets the last valid capability on each call to
SetCapsForExec(). An alternative implementation would turn 'CapSet'
into a full-fledged class, and get the last valid capability in the
constructor. I went with the current implementation to avoid having
~40 syscalls run in the constructor.

Bug: 32438163
Test: Use a test service that uses capabilities.
Test: Apply in internal tree and test with angler and rild.

Change-Id: Ia271cc7eb389d1d526d61f897261e4bac4d19e5d