liblog: Add internal __android_log_ratelimit_identical
In the liblog properties, there is a caching algorithm that
uses the bionic private interfaces to clearly understands if
the properties are dirty and needs to reload.  Typically 22ns
lookup, rather than 500ns for a regular call using the public
bionic interface.  We do so whenever a property is in the path
of logging.

This new internal property handler __android_log_ratelimit_identical
provides a fast answer to first character converted to lowercase of
logd.ratelimit.identical property.  The goal is to accept commands
like "sleep" or "DROP" and simply select on the first character.

Test: compile
Bug: 33535908
Change-Id: Ib07024988e4c44e59ae24659970aa8e1357c55d5