liblog: Add internal __android_log_ratelimit_identical
In the liblog properties, there is a caching algorithm that uses
bionic private interfaces to clearly understand if the properties are
dirty and needs to reload.  This typically takes 22ns to perform the
lookup, rather than 500ns for a regular call using the public
bionic interface.  We do so whenever a property is in the path
of logging.

This new internal property handler __android_log_ratelimit_identical
implements a fast lookup of logd.ratelimit.identical, and a conversion
of the first character to lowercase for caller convenience.  The goal
is to accept commands like "sleep" or "DROP" and simply select on the
's' or 'd'.

Test: compile
Bug: 33535908
Change-Id: Ib07024988e4c44e59ae24659970aa8e1357c55d5