liblog: Add internal __android_log_timestamp_ratelimit
Tell us if the timestamps that were submitted over the time horizon
exceed the maximum number of logging incidents.  Are we submitting
logs too fast or not?  This is global function in the process,
maintaining internal state, but to recognize that we could be called
in a signal handler context, we return -1 if the internal trylock
request fails.

Test: compile
Bug: 33535908
Change-Id: I9a76f727bda0b72d37f3b46e8a161232ece17656