Update patch set 1
Patch Set 1: Verified+1 Presubmit-Ready+1

Explanation:

limiting uid, pid or tid may be bad in the first place, please not up for debate for pragmatic performance reasons. This deals with the recent restructuring of the base uid map to beyond 40000.

A followup CL may try to extend the key to support full range for all process values, but logging has limited pid and tid to uint16_t when transmitted over the wire so no gain would result to use uid_t and pid_t in the map.

Patch-set: 1
Label: Verified=+1