Update patch set 3
Patch Set 3:

> Looks OK to me :
 > 
 > @enh, it wasn't clear to me on code search whether the preferred
 > idiom was to use a compiler flag (-D_FILE..) or to #define
 > _FILE_OFFSET_BITS within a given translation unit. I also didn't
 > spend much time weighing the pros and cons of the two approaches.

-D (at least when the makefile is simple) seems less error-prone; it's harder to end up with one file built one way and another file another. i think man page examples tend to use #define just because they don't use makefiles :-)

i checked the current .h files, and they only have one reference to off_t and explicitly say off64_t.

the only catch here is that we need to be careful not to add an off_t to those headers in future, at least in C API. it might be worth removing the __BEGIN_DECLS/__END_DECLS from the headers since i _think_ the headers are already C++-only, and i _think_ the callers are already C++-only too. i'll upload that change to find out...

Patch-set: 3