Allow system properties having @ and ! characters
This change modifies the property_service of the init process so that
system properties having @ and ! characters are allowed.

Also this change extends the simple protocol for reqesuting an update
for a system property: prop_msg. Depending on the MSB of the cmd field
in prop_msg, two different versions of prop_msg are used accordingly.

Bug: 33926793
Test: setprop and getprop command with very long name and value
(up to 2^9-1) should succeed.

Change-Id: Ib04cb3b2cb69efada3b2d20c854ffb29debeaf0f