WIP: Allow longer values and names for sysprops
This change modifies the protocol between the init process and client
process requesting an update of a sysprop. In order to allow longer name
and value for a sysprop, the prop_msg structure has also been enlarged.

This change is just to accomodate the enlarged prop_msg structure and to
be consistent with the new limits on name/value.

Bug: 33926793
Test: setprop and getprop command with very long name and value
(up to 2^9-1) should succeed.

Change-Id: Ib04cb3b2cb69efada3b2d20c854ffb29debeaf0f