Update patch set 2
Patch Set 2:

Orthogonally, I am afraid that logd running at background cgroup was the problem, the logger (used to be) DOS'd when the rate was higher than 20/second. I believe 20/second (which is really 60/second when split out) is a now a relatively safe value except for the fact it exceeds the sustained 5/second rule that defines the difference between a spammy application and not. We could add a separate CL to spot greater than 50 / ten seconds and roll back the ratelimit. I was not aware I could dynamically control it ...

Patch-set: 2