Update patch set 3
Patch Set 3:

nnk@ I do want your input on the tuneables so as to not duly impact the usefulness of auditd. For instance we could probably increase the maximum burst rate from 20 to something you might be happier with? Or drop the maximum sustained rate to something lower than 5 if you know for sure that it would have a positive impact on power consumption under sepolicy (DOS?) scenarios?

Patch-set: 3