Update patch set 3
Patch Set 3:

Forgot to note above that EINTR is already managed with a retry loop.

Chickened out once this code inspected, switching to void also means we assume errno is not affected, which it is ... it is an anti-pattern to not have an int return in such a case (POSIX has even defined close() with one despite it _never_ being used). It would be additional code complexity to preserve errno ...

Patch-set: 3