Update patch set 3
Patch Set 3:

SafetyNet contains two variable length strings, so one can _not_ create a SafteyNet structure.  We do not wish to use the new C++ events log decoder because it is the same code that composes the messages and thus doesn't test the events for correct binary layout. Let the gTest code for SafetyNet generation stand as it independent confirmation of binary identity ...

Patch-set: 3