Update patch set 6
Patch Set 6: Code-Review-2

We talked this over a bit, and we've come around to the idea that we probably shouldn't be including this API in the NDK. Its heavy use of macros is, as we've seen, pretty invasive and can cause some problems.

enh did some searching and it looks like there are a handful of third-party logging libraries that are easy to use in the NDK. We haven't heard a single report of people wishing this full API was available (it was just us thinking that we should), so I think avoiding churn is the best bet until we hear otherwise.

Patch-set: 6
Label: Code-Review=-2