Update patch set 1
Patch Set 1:

(2 comments)

I agree with the concept here that we should clean up the return value of the builtins.  

I think if we're going to invest effort into making this change that we can do better than just replacing the return value with bool.  

There are plenty of builtins that call multiple functions, each of which can fail, setting errno and/or a return value, but there's no  visibility into these failures.  It would be good to percolate these out to the existing INFO print in ExecuteCommand() and we could use the std::string* err parameter pattern that's already used throughout init to do so.  We could then always print this 'command' string in the case of command failure in addition to when we're logging at the info level.

Thoughts?

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0