Update patch set 1
Patch Set 1:

I've realized that the kernel prebuilt for marlin actually comes from the android-msm-marlin-3.18-nougat-mr1 kernel branch, and may not contain all the patches we need for endpoint_alloc. Also we're unable to actually update kernels in aosp. I think its better to submit just this patch and see improvements on master once it automerges, and update the properties in aosp when the prebuilts are updated.

Patch-set: 1