Update patch set 2
Patch Set 2:

I should split this between a local logger with writer/reader interfaces, that could be landed in host as a fake logger for testing purposes. Provide an 'enabler' call regarding the ToDo item for permissions. Strip out C binder interfaces (also in a draft review), etc so that this code can be run on host or client.

Patch-set: 2
Label: Verified=0