Update patch set 1
Patch Set 1:

The larger r/w is somewhat dependent on EPFILE_ALLOC working as intended. the ioctl will allocate a buffer so that every syscall doesn't allocate and free memory, which can cause fragmentation and eventually lead to ENOMEM. This can cause crashes with large files. I haven't had a problem with adb in this cause because the r/w size is actually limited to 65kB by the protocol, but MTP has crashed with 255kB r/w size and I don't want to risk it in this case. We're also not allowed to do kernel drops in aosp, the device team handles them at particular times.

Patch-set: 1