Don't undef __unused if we're using Bionic
__unused is used by Bionic as __attribute__((unused)), but some of the
glibc headers in prebuilts/ use it as a variable name. It would be nice
if we could use __unused in fcntl (and any files included by fcntl), so
we shouldn't undef if if we're using Bionic.

This undef is only an issue if we've already included Bionic's
sys/cdefs.h, which is what defines __BIONIC__. So, it shouldn't be an
issue if fcntl.h is the first Bionic header we're including.

Bug: 32073964
Test: Can now use __unused in bits/fcntl without compiler errors.
Change-Id: Ic337d3a22bc73620ad42f8a5f5528c163b6f2702