Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

only logd.klogd is allowed to submit to the kernel log buffer.

You did find a bug though, not a serious one because if someone tries to log to the LOG_ID_KERNEL from user space it will be dropped on the floor by logd.

Patch-set: 1
Label: Code-Review=-1