fs_mgr: Add support for at-boot verification in AVB
Running dm-verity on heavily accessed partitions leads to performance
slowdowns, especially on low-RAM and slow-CPU devices.

This patch checks the fs_mgr flag 'verityatboot': to allow an entire verified
partition to be read once at boot, to check for corruptions. If the reads are
successful, we can mount the partition as raw & read-only, and if not,
we can revert to mounting it as a verity partition, just like before.

Usage of this flag will entail a slowdown of time-to-boot, but should
lead to improvements in runtime performance.

Bug: 34779406
Test: Device can boot
Change-Id: I08409a394278e89042f04f648962feeab78b7465