Update patch set 2
Patch Set 2:

It does break sendevent, but people really shouldn't be using it for automated testing at this point. Pretty much everyone should have already moved to UiAutomation#injectInputEvent if they're running instrumentation tests (i.e. CTS), Monkey for their stress tests, and the input command (`adb shell input ...`) for injecting swipes and things. The only thing this really breaks is exactly reproducing user input as if it came from the attached input device... which is kind of the point. The abuse team is seeing a lot of attacks that use this to masquerade as a real user.

Anyways, based on a quick search only one team internally is using it, and that's only because we didn't support setting the duration for a swipe until L. I'll reach out to them to mention this is going to break in O.

Patch-set: 2
Reviewer: Gerrit User 1006497 <1006497@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0