Update patch set 8
Patch Set 8: Verified+1 Presubmit-Ready+1

Passes all existing tests, the tests that matter at this moment.

ToDo: The bocking tests are excluded since this functionality is to be added later. In my private branch I have a WIP prototype for SafetyNet messages going to the security buffer, and it fails liblog_local.android_errorWriteWithInfoLog__android_logger_list_read__security which stresses the writer and readers as these messages are to go to logd, and there are multiple sources for the content (local for everything else, logd for security buffer). We will consider fixing this in a future patch as well. liblog.android_errorWriteWithInfoLog__android_logger_list_read__security passes in the simpler case where local logger is not present, so the fix should(?) be in the common code, and not the transports.

Patch-set: 8